refactor: fix typo in file names
